k8s refactor
